refactor(gh): unify frontend diagnostics markdown rendering
